Added a simple scene, removed the buttons from the Holster, changed POS of EndDialog
Added a simple scene to the view so that it looks as if we are playing a game. Also I removed the buttons from the under Dash Foods and added them to their Own Canvas. This way they are not manipulated when you move the Food Order canvas, this was causing issue with the button getting in the way and was just not looking right(they are now stationary). Adjusted the particle system to the button. Moved Dialog Box to be centered in the Holster bounding control box.